# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: فزعة FAZAAH – منصة الخدمات والمهنيين في اليمن
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersPage"
  /admin/users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [admin]
      summary: Update user status (ban/activate)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
  /admin/providers/{id}/verify:
    patch:
      operationId: verifyProvider
      tags: [admin]
      summary: Verify/reject provider account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderVerifyInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
...
  /admin/subscription-payments/{id}/review:
    patch:
      operationId: reviewSubscriptionPayment
      tags: [admin]
      summary: Approve or reject a subscription payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentReviewInput"
      responses:
        "200":
          description: Reviewed payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionPayment"
  /admin/advertisements/{id}/review:
    patch:
      operationId: reviewAdvertisement
      tags: [admin]
      summary: Approve or reject an advertisement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvertisementReviewInput"
      responses:
        "200":
          description: Reviewed advertisement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Advertisement"

  # ─── WOW: HOME FEED ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SubscriptionPayment:
      type: object
      required: [id, providerId, plan, wallet, transactionReference, status, createdAt]
      properties:
        id:
          type: integer
        providerId:
          type: integer
        plan:
          type: string
          enum: [monthly, yearly]
        wallet:
          type: string
          enum: [jeeb, floosk, jawali, cash, one_cash, hasib, easy]
        transactionReference:
          type: string
        receiptUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, expired, refunded]
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
        reviewedAt:
          type: ["string", "null"]
...
    PaymentReviewInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        adminNote:
          type: ["string", "null"]
...
    Advertisement:
      type: object
      required: [id, providerId, title, description, city, district, plan, durationDays, budget, status, createdAt]
      properties:
        id:
          type: integer
        providerId:
          type: integer
        providerName:
          type: ["string", "null"]
        categoryName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
        city:
          type: string
        district:
          type: string
        targetAudience:
          type: ["string", "null"]
        plan:
          type: string
          enum: [standard, featured, homepage]
        durationDays:
          type: integer
        budget:
          type: number
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, rejected, expired]
        reviewNote:
          type: ["string", "null"]
        startsAt:
          type: ["string", "null"]
        endsAt:
          type: ["string", "null"]
        createdAt:
          type: string
...
    AdvertisementReviewInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [active, rejected]
        reviewNote:
          type: ["string", "null"]

    # ── AUTH ────────────────────────────────────────────────────────────────────
...
    SuccessResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # ── ADMIN ───────────────────────────────────────────────────────────────────
...
    AdminUsersPage:
      type: object
      required: [users, total, page]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: integer
    AdminUser:
      type: object
      required: [id, name, phone, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
        status:
          type: string
        avatarUrl:
          type: ["string", "null"]
        categoryName:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        completedJobs:
          type: ["integer", "null"]
        isVerified:
          type: ["boolean", "null"]
        createdAt:
          type: string
    UserStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [active, banned, pending]
    ProviderVerifyInput:
      type: object
      required: [isVerified]
      properties:
        isVerified:
          type: boolean
        rejectionReason:
          type: ["string", "null"]
...