# lib/api-zod/src/generated/types/paymentReviewInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * فزعة FAZAAH – منصة الخدمات والمهنيين في اليمن
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentReviewInputStatus } from './paymentReviewInputStatus';

export interface PaymentReviewInput {
  status: PaymentReviewInputStatus;
  /** @nullable */
  adminNote?: string | null;
}