# lib/api-zod/src/generated/types/subscriptionPayment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * فزعة FAZAAH – منصة الخدمات والمهنيين في اليمن
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionPaymentPlan } from './subscriptionPaymentPlan';
import type { SubscriptionPaymentStatus } from './subscriptionPaymentStatus';
import type { SubscriptionPaymentWallet } from './subscriptionPaymentWallet';

export interface SubscriptionPayment {
  id: number;
  providerId: number;
  plan: SubscriptionPaymentPlan;
  wallet: SubscriptionPaymentWallet;
  transactionReference: string;
  /** @nullable */
  receiptUrl?: string | null;
  status: SubscriptionPaymentStatus;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
  /** @nullable */
  reviewedAt?: string | null;
}